+++ lib/src/inherited_provider.dart
@@ -1,4 +1,3 @@
-
 part of 'inherited_object.dart';
 
 
@@ -125,4 +124,21 @@ class InheritedDataProviderState<T>
   }
 }
 
+
+///
+///
+///
+class InheritedProviders extends StatelessWidget {
+  final List<InheritedObjectProvider> entries;
+  final Widget child;
+  const InheritedProviders(this.entries,{super.key, required this.child});
+
+  @override
+  Widget build(BuildContext context) {
+    return entries.isEmpty ? child : entries.reversed.skip(1).fold(
+      entries.last.copyWithChild(child),
+      (previous, current) => current.copyWithChild(previous)
+    ); // Chain the providers and return the final widget
+  }
+}